feat: add 'kind' parameter to Reporter$image() and allow to pass extra arguments to the report
Add a new parameter 'kind' to the 'image' method of the Reporter class. This parameter allows specifying the kind of report, with the default value set to "image". The 'kind' parameter is added to the output list of the report, along with any other additional arguments passed to the method.

Example usage:
reporter <- get_reporter()
reporter$add(
   list(
      name = "Image 1",
      contents = list(reporter$image("/path/to/image1", "pdf", save_code = TRUE, kind = "table_image", descr = "blah")),
   ),
   h1 = "Images",
   h2 = "Image 1"
)